+++ migration/sql-files/047-codebases.sql
@@ -3,7 +3,7 @@ CREATE TABLE codebases (
     updated_at timestamp with time zone,
     deleted_at timestamp with time zone,
     id uuid primary key DEFAULT uuid_generate_v4() NOT NULL,
-    space_id uuid,
+    space_id uuid NOT NULL REFERENCES spaces (id) ON DELETE CASCADE,
     type text,
     url text
 );
